Allow maxAge parameter on Clamo requests
Clamo does not implement standard HTTP Cache-Control.

Instead, by default, GETs are cached for 180 seconds in Varnish
controlled by a combination of an X-Header and the URL.

This causes a problem for applications fetching content as the
expectation is that fastFT should publish fresh content within a few
seconds of it being available in the API.

We can work around this problem by specifying a maxAge parameter, that
in effect acts as a cache buster. The parameter contains a integer that
is a timestamp to the nearest n seconds.